move state store impl to separated file
